Basic implementation of mmo
See DESIGN.md and SIMPLE_IMPL.md for the idea behind this.

Auth is implemented as query parameters in the mcp connection url.
The world is persisted in memory.

Players can be autonomous or they can be human, depending on whether a
player is autonomous or not the prompts change slightly. (to encourage
the llm to play by itself or assist a human).

Rooms are different locations that players can move between.
Players can take actions and leave lingering effects on each other /
rooms.

The last major missing piece is items.