Adding passwords
Adding a password that we authenticate in the url. This is not great,
but minimizes friction (oauth seems so heavy handed). Given this is just
a game, it'll due. We picked the number of bcrypt rounds to be fast, so
it's not really that secure. But again, it's jsut a game.